refactor: abstract past recipients into seperate component
